modify Board entity

+++ src/main/java/com/fastcampus/board/entity/Board.java
@@ -1,6 +1,5 @@
 package com.fastcampus.board.entity;
 
-import com.fastcampus.board.repository.BoardRepository;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,8 +9,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Data
 @AllArgsConstructor
@@ -21,15 +18,11 @@ public class Board {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer seq;
+    private long seq;
 
     private String title;
-
     private String writer;
-
     private String content;
-
     private LocalDateTime regDate;
-
     private Integer cnt;
 }
